backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Pillow==10.0.1
python-dotenv==1.0.0
bleach==6.1.0
PyJWT==2.8.0
redis==5.0.1
bcrypt==4.1.2
marshmallow==3.20.1
